PathFinder/requirements.txt
rpi.gpio==0.6.3
future==1.0.0
iso8601==2.1.0
numpy==1.26.4
opencv-python==4.9.0.80
pygame==2.5.2
pyserial==3.5
PyYAML==6.0.1
